new code

+++ app/src/main/java/com/phm/myhello/model/Amount.kt
@@ -1,7 +1,7 @@
 package com.phm.myhello.model
 
 data class Amount(
-        var id: Long
+        var id: Int
         , var date: Int
         , var month: Int
         , var year: Int
